scene_gen/requirements.txt
h5py
numpy
pyyaml
pytorch-fid